Move exceptions and json patcher into another modules
