+++ rss_reader/utils/exceptions.py
@@ -1,0 +1,22 @@
+"""Custom Rss Exceptions"""
+
+
+class RssException(Exception):
+    """
+    Custom Exception class raised by RssBots classes
+    """
+    pass
+
+
+class RssValueException(ValueError):
+    """
+    Custom Exception class raised by RssBots classes
+    """
+    pass
+
+
+class RssNewsException(ValueError):
+    """
+    Custom Exception class raised by RssBots classes
+    """
+    pass
